lib/src/test/java/com/bridgelab/hotelreservationsystem/HotelReservationTest.java
package com.bridgelab.hotelreservationsystem;
import static org.junit.Assert.assertEquals;
import java.util.Arrays;
import java.util.List;

import org.junit.Test;

public class HotelReservationTest {
	@Test
	public void whenHotelAddedToSystemShouldGetAdded()
	{
		Hotel hotel1 = new Hotel("Lakeewood", 110);
		Hotel hotel2 = new Hotel("Bridgewood", 160);
		Hotel hotel3 = new Hotel("Ridgewood", 220);
		Hotel[] hotelList = { hotel1, hotel2, hotel3 };
		List<Hotel> hotels = Arrays.asList(hotelList);
		HotelReservation hotelReservation = new HotelReservation();
		hotelReservation.add(hotel1);
		hotelReservation.add(hotel2);
		hotelReservation.add(hotel3);
		List<Hotel> result = hotelReservation.getHotelList();
		assertEquals(hotels.size(), result.size());
	}
}
